% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bmi.bmi}
\alias{bmi.bmi}
\alias{hdl.hdl}
\alias{bmi.hcl}
\alias{abo.chd}
\alias{vtn.chd}
\alias{fcgr2a.chd}
\title{Example datasets}
\format{
An object of class \code{data.frame} with 1778 rows and 29 columns.

An object of class \code{data.frame} with 1782 rows and 29 columns.

An object of class \code{data.frame} with 1798 rows and 29 columns.

An object of class \code{data.frame} with 2009 rows and 29 columns.

An object of class \code{data.frame} with 2008 rows and 29 columns.

An object of class \code{data.frame} with 2005 rows and 29 columns.
}
\source{
\url{https://gwas.mrcieu.ac.uk/datasets/}
}
\usage{
bmi.bmi

hdl.hdl

bmi.hcl

abo.chd

vtn.chd

fcgr2a.chd
}
\description{
The example datasets prepared for MR-GMM:
\itemize{
\item \code{bmi.bmi}: BMI and BMI;
\item \code{hdl.hdl}: HDL and HDL;
\item \code{bmi.hcl}: BMI and Hair color (Blond);
\item \code{abo.chd}: Histo-blood group ABO system transferase and CHD;
\item \code{vtn.chd}: Vitronectin and CHD;
\item \code{fcgr2a.chd}:  Low affinity immunoglobulin gamma Fc region receptor II-a and CHD.
}
}
\keyword{datasets}
