% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{mr.gmm}
\alias{mr.gmm}
\title{Conduct MR-GMM method}
\usage{
mr.gmm(
  beta_exp,
  beta_out,
  se_exp,
  se_out,
  sel.p = 1,
  inits = NULL,
  lambda = NULL,
  DA = T,
  xi = ifelse(DA, 0.1, 1),
  rate = ifelse(DA, 1.1, 1),
  maxit = 10000,
  tol = 1e-07
)
}
\arguments{
\item{beta_exp}{An m-dimensional vector of effect sizes from the exposure GWAS data.}

\item{beta_out}{An m-dimensional vector of effect sizes from the outcome GWAS data.}

\item{se_exp}{An m-dimensional vector of standard errors for \code{beta_exp} from the exposure GWAS data.}

\item{se_out}{An m-dimensional vector of standard errors for \code{beta_out} from the outcome GWAS data.}

\item{sel.p}{The p-value threshold for selecting SNPs. Default is 1.}

\item{inits}{A list containing the initial parameter estimates for the Gaussian Mixture Model (GMM).
The list must include four elements: \code{beta}, \code{tau}, \code{eta}, and \code{pi}.
\itemize{
\item \code{beta}: The causal effect.
\item \code{tau}: The inverse variance of latent pleiotropic effects on the outcome.
\item \code{eta}: The inverse variance of latent SNP effects on the exposure.
\item \code{pi}: The probabilities for the four clusters: invalid, valid, invalid & null, and null IVs.
The default is a random start.
}}

\item{lambda}{The penalty on \code{pi}. Default is the square root of m.}

\item{DA}{A logical value indicating whether the deterministic annealing procedure should be applied. Default is TRUE.}

\item{xi}{The initial temperature for the deterministic annealing (DA) procedure.
If \code{DA = TRUE}, \code{xi} defaults to 0.1; otherwise, it is set to 1.}

\item{rate}{Temperature scheduling parameter. In each iteration, \code{xi} is increased by multiplying \code{rate}. Default is 1.1.}

\item{maxit}{Maximum number of iterations of the (DA)EM algorithm. Default is 10,000.}

\item{tol}{The convergence tolerance. Default is 1e-7.}
}
\value{
A list containing the following components:
\itemize{
\item \code{data} An m × 4 data frame of the input GWAS data.
\item \code{beta.hat} The causal effect estimate.
\item \code{beta.se} The standard error of \code{beta_hat}.
\item \code{beta.p.value} The p-value from the two-tailed Wald test for \eqn{H_0: \beta=0}.
\item \code{tau.hat} The estimated inverse variance of latent pleiotropic effects on the outcome.
\item \code{eta.hat} The estimated inverse variance of latent SNP effects on the exposure.
\item \code{rho} An m × 4 matrix indicating the estimated latent probabilities of each SNP belonging to the four clusters.
\item \code{pi} The estimated probabilities for the four clusters.
\item \code{ELBO} A vector recording the Evidence Lower BOund (ELBO) at each iteration, useful for assessing convergence.
\item \code{cluster} An m × 2 data frame. The first column contains the most probable cluster assignments for each SNP, and the second column contains the corresponding estimated latent probabilities.
\item \code{iter} The total number of iterations performed by the DAEM algorithm.
}
}
\description{
Main function for MR-GMM
}
\examples{
data(bmi.bmi)
attach(bmi.bmi)
bmi.bmi.res <- mr.gmm(beta.exposure, beta.outcome, se.exposure, se.outcome)
plot_wald(bmi.bmi.res)
plot_scatter(bmi.bmi.res)

}
\references{
Yuquan Wang, Yunlong Cao, Dong Chen, Dapeng Shi, Liwan Fu, et al. Pseudo-p-Value-Based Clumping Enhanced Proteome-wide Mendelian Randomization with Application in Identifying Coronary Heart Disease-Associated Plasma Proteins. \url{https://doi.org/10.1101/2025.01.13.25320450}.
}
