% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_wald}
\alias{plot_wald}
\alias{plot_scatter}
\title{Visualize the results of MR-GMM}
\usage{
plot_wald(res)

plot_scatter(res)
}
\arguments{
\item{res}{A list obtained from \code{mr.gmm}.}
}
\value{
R plots.
}
\description{
Visualization functions for MR-GMM analysis. Note that the clusters shown in the plots represent the most probable clusters of SNPs, not the true clusters.
\itemize{
\item \code{plot_wald}: Plots the Wald ratio estimates against IV strength.
\item \code{plot_scatter}: Plots the IV-outcome associations against IV-exposure associations.
}
}
